Kidobó módosítása xml beolvasásra Verzió léptetés

+++ V_Ablakok/6_Kiadási_adatok/Kidobó/Ablak_kidobó.cs
@@ -5,7 +5,6 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Windows.Forms;
-using System.Xml.Linq;
 using Villamos.Adatszerkezet;
 using Villamos.Kezelők;
 using Villamos.V_MindenEgyéb;
@@ -237,50 +236,23 @@ namespace Villamos
                 else
                     return;
 
-                //Kísérlet
-
-                XDocument doc = XDocument.Load(fájlexc);
-
-                var allNames = doc.Descendants()
-                                  .Select(a => a.Name.LocalName)
-                                  .Distinct()
-                                  .OrderBy(n => n)
-                                  .ToList();
 
 
-                foreach (var name in allNames)
-                    szöveg += $"- {name}";
-
-
-
-
-
-
-
-                //Kísérlet
-
-
-                // megnyitjuk a beolvasandó táblát
-                string munkalap = "Munka1";
-                MyX.ExcelMegnyitás(fájlexc);
-
-                // leellenőrizzük, hogy az adat nap egyezik-e
-                szöveg = MyX.Beolvas(munkalap, "a2").Trim().Replace(".", "");
-                if (MyX.Beolvas(munkalap, "a2").Trim().Replace(".", "") != Dátum.Value.ToString("yyyyMMdd"))
+                DateTime MelyikNap = MyX.KidobóDátumEll(fájlexc);
+                if (MelyikNap.ToShortDateString() != Dátum.Value.ToShortDateString())
                 {
                     // ha nem egyezik akkor       
-                    // az excel tábla bezárása
-                    MyX.ExcelBezárás();
                     Holtart.Ki();
                     throw new HibásBevittAdat("A betölteni kívánt adatok nem egyeznek meg a beállított nappal !");
                 }
+
+
                 string hely = $@"{Application.StartupPath}\{Cmbtelephely.Text.Trim()}\Adatok\Főkönyv\Kidobó\{Dátum.Value.Year}\{Dátum.Value:yyyyMMdd}Forte.mdb";
                 if (File.Exists(hely))
                 {
                     if (MessageBox.Show("Már van az adott napra feltöltve adat ! Módosítjuk az adatokat ?", "Figyelmeztetés", MessageBoxButtons.YesNo, MessageBoxIcon.Information) == DialogResult.No)
                     {
                         // Nemet választottuk
-                        MyX.ExcelBezárás();
                         Holtart.Ki();
                         return;
                     }
@@ -291,52 +263,20 @@ namespace Villamos
                         File.Delete(hely);
                     }
                 }
-                // megnézzük, hogy hány sorból áll a tábla
-                int ii = 4;
-                int utolsó = 0;
 
-                while (MyX.Beolvas(munkalap, $"a{ii}").Trim() != "_")
+                List<string> hibak;
+                List<Adat_Kidobó> adatok = MyX.BeolvasKidobo(fájlexc, out hibak);
+                if (hibak.Count > 0)
                 {
-                    utolsó = ii;
-                    ii += 1;
-                    szöveg = MyX.Beolvas(munkalap, $"a{ii}").Trim();
+                    foreach (var h in hibak)
+                        szöveg += $"Hiba: {h}/n";
+                    throw new Exception(szöveg);
                 }
 
-                Holtart.Be(utolsó + 1);
-                if (utolsó > 1)
-                {
-                    // megnyitjuk a táblát
-                    List<Adat_Kidobó> Adatok = new List<Adat_Kidobó>();
-                    for (int i = 5; i <= utolsó; i++)
-                    {
-                        string ideig = MyX.Beolvas(munkalap, $"a{i}");
-                        string[] darabol = ideig.Split('/');
-                        string viszonylat = darabol[0].Trim();
-                        Adat_Kidobó Adat = new Adat_Kidobó(
-                            MyF.Szöveg_Tisztítás(viszonylat, 0, 6),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"b{i}").Trim(), 0, 6),
-                            MyF.Szöveg_Tisztítás(ideig.Trim(), 0, 20),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"d{i}").Trim(), 0, 100),
-                            MyX.Beolvasidő(munkalap, $"f{i}"),
-                            MyX.Beolvasidő(munkalap, $"h{i}"),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"g{i}").Trim(), 0, 50),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"i{i}").Trim(), 0, 50),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"e{i}").Trim(), 0, 3),
-                            "_",
-                            "_",
-                            "_",
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"k{i}").Trim(), 0, 30),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"c{i}").Trim(), 0, 10));
+                KézKidobó.Rögzítés(Cmbtelephely.Text.Trim(), Dátum.Value, adatok);
 
-                        Adatok.Add(Adat);
-                        Holtart.Lép();
-                    }
-
-                    KézKidobó.Rögzítés(Cmbtelephely.Text.Trim(), Dátum.Value, Adatok);
-                }
                 Gombok();
 
-                MyX.ExcelBezárás();
                 // kitöröljük a betöltött fájlt
                 File.Delete(fájlexc);
 
@@ -361,6 +301,7 @@ namespace Villamos
 
         private void Forte_Beolvasás_Click(object sender, EventArgs e)
         {
+            string szöveg = "";
             string fájlexc = "";
             try
             {
@@ -379,76 +320,32 @@ namespace Villamos
                 else
                     return;
 
-
-                // megnyitjuk a beolvasandó táblát
-                string munkalap = "Munka1";
-                MyX.ExcelMegnyitás(fájlexc);
-
-                string hely = $@"{Application.StartupPath}\{Cmbtelephely.Text.Trim()}\Adatok\Főkönyv\Kidobó\{Dátum.Value.Year}\{Dátum.Value:yyyyMMdd}Forte.mdb";
-
-                // leellenőrizzük, hogy az adat nap egyezik-e
-                string szöveg = MyX.Beolvas(munkalap, "a2").Trim().Replace(".", "");
-                if (MyX.Beolvas(munkalap, "a2").Trim().Replace(".", "") != Dátum.Value.ToString("yyyyMMdd"))
+                DateTime MelyikNap = MyX.KidobóDátumEll(fájlexc);
+                if (MelyikNap.ToShortDateString() != Dátum.Value.ToShortDateString())
                 {
-                    // ha nem egyezik akkor
-                    MessageBox.Show("A betölteni kívánt adatok nem egyeznek meg a beállított nappal !", "Figyelmeztetés", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    // az excel tábla bezárása
-                    MyX.ExcelBezárás();
+                    // ha nem egyezik akkor       
                     Holtart.Ki();
-                    return;
+                    throw new HibásBevittAdat("A betölteni kívánt adatok nem egyeznek meg a beállított nappal !");
                 }
 
+                string hely = $@"{Application.StartupPath}\{Cmbtelephely.Text.Trim()}\Adatok\Főkönyv\Kidobó\{Dátum.Value.Year}\{Dátum.Value:yyyyMMdd}Forte.mdb";
                 if (!File.Exists(hely))
                 {
                     MessageBox.Show("A választott napra még nincs feltöltve adat ! ", "Figyelmeztetés", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    MyX.ExcelBezárás();
                     Holtart.Ki();
                     return;
                 }
 
-
-                // megnézzük, hogy hány sorból áll a tábla
-                int ii = 4;
-                int utolsó = 0;
-
-                while (MyX.Beolvas(munkalap, $"a{ii}").Trim() != "_")
+                List<string> hibak;
+                List<Adat_Kidobó> adatok = MyX.BeolvasKidobo(fájlexc, out hibak);
+                if (hibak.Count > 0)
                 {
-                    utolsó = ii;
-                    ii += 1;
-                    szöveg = MyX.Beolvas(munkalap, $"a{ii}");
+                    foreach (var h in hibak)
+                        szöveg += $"Hiba: {h}/n";
+                    throw new Exception(szöveg);
                 }
 
-                Holtart.Be(utolsó + 1);
-                if (utolsó > 1)
-                {
-                    // megnyitjuk a táblát
-                    List<Adat_Kidobó> Adatok = new List<Adat_Kidobó>();
-                    for (int i = 5; i <= utolsó; i++)
-                    {
-                        string ideig = MyX.Beolvas(munkalap, $"a{i}");
-                        string[] darabol = ideig.Split('/');
-                        string viszonylat = darabol[0].Trim();
-                        Adat_Kidobó Adat = new Adat_Kidobó(
-                            MyF.Szöveg_Tisztítás(viszonylat, 0, 6),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"b{i}").Trim(), 0, 6),
-                            MyF.Szöveg_Tisztítás(ideig.Trim(), 0, 20),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"d{i}").Trim(), 0, 100),
-                            MyX.Beolvasidő(munkalap, $"f{i}"),
-                            MyX.Beolvasidő(munkalap, $"h{i}"),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"g{i}").Trim(), 0, 50),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"i{i}").Trim(), 0, 50),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"e{i}").Trim(), 0, 3),
-                            "_",
-                            "_",
-                            "_",
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"k{i}").Trim(), 0, 30),
-                            MyF.Szöveg_Tisztítás(MyX.Beolvas(munkalap, $"c{i}").Trim(), 0, 10));
-
-                        Adatok.Add(Adat);
-                        Holtart.Lép();
-                    }
-                    KézKidobó.Rögzítés(Cmbtelephely.Text.Trim(), Dátum.Value, Adatok);
-                }
+                KézKidobó.Rögzítés(Cmbtelephely.Text.Trim(), Dátum.Value, adatok);
 
                 Gombok();
